Enhance precision of longitude and latitude data by changing their type from float to double.
Reason: http://wiki.openstreetmap.org/wiki/Node - "Do not use IEEE 32-bit floating point data type since it is limited to about 5 decimal places."

+++ src/OsmSharp/Db/HistoryDb.cs
@@ -174,7 +174,7 @@ namespace OsmSharp.Db
         /// <summary>
         /// Gets all visible objects within the given bounding box.
         /// </summary>
-        public IEnumerable<OsmGeo> Get(float minLatitude, float minLongitude, float maxLatitude, float maxLongitude)
+        public IEnumerable<OsmGeo> Get(double minLatitude, double minLongitude, double maxLatitude, double maxLongitude)
         {
             return _db.Get(minLatitude, minLongitude, maxLatitude, maxLongitude);
         }
